+++ ui-plugins/muse-layout-antd/craco.config.js
@@ -1,18 +1,16 @@
 const CracoLessPlugin = require('craco-less');
 const MuseCracoPlugin = require('@ebay/muse-craco-plugin');
-const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+const MuseEbayCracoPlugin = require('@ebay/muse-ebay-craco-plugin');
 
 module.exports = () => {
   return {
-    plugins: [{ plugin: CracoLessPlugin }, { plugin: MuseCracoPlugin }],
-    webpack: {
-      plugins: {
-        add: [new BundleAnalyzerPlugin({ analyzerMode: 'static', openAnalyzer: false })],
-      },
-    },
+    plugins: [
+      { plugin: CracoLessPlugin },
+      { plugin: MuseCracoPlugin },
+      { plugin: MuseEbayCracoPlugin },
+    ],
     jest: {
       configure: {
-        // override default jest configuration provided by @ebay/muse-craco-plugin
         setupFiles: ['<rootDir>/tests/setup.js'],
         testMatch: ['<rootDir>/tests/**/*.test.js'],
         roots: ['<rootDir>/tests/'],